+++ apps/dashboard/src/routes/admin/vendors/_index.tsx
@@ -29,7 +29,7 @@ export default function VendorsPage() {
         </Button>
       </div>
 
-      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm'>
+      <section className='rounded-lg bg-card p-6 text-card-foreground shadow-sm dark:border'>
         <Table>
           <TableHeader>
             <TableRow>